services/ai_resource_recognition/requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://bytedpypi.byted.org/simple

# ---- version guards ----
numpy<2

# ---- pytorch stack (cu121) ----
torch==2.1.0+cu121
torchvision==0.16.0+cu121

# ---- faiss (lock to CUDA12-compatible) ----
faiss-gpu==1.7.2

# ---- others ----
tqdm>=4.66
pyarrow>=14.0.0
pillow>=10.0.0
omegaconf>=2.3.0
torchmetrics==1.2.1
opencv-python==4.9.0
imagehash
bytedtos
volcengine-python-sdk[ark]

byted_overpass_toutiao_article_article
bytedeuler==2.6.4